Transaction History

+++ src/main/java/org/BankingSystem/models/TransactionHistory.java
@@ -3,10 +3,10 @@ package org.BankingSystem.models;
 public class TransactionHistory {
     String transactionID;
     String date;
-    int amount;
+    float amount;
     TypeOfTransaction type;
 
-    public TransactionHistory(String transactionID, String date, int amount, TypeOfTransaction type) {
+    public TransactionHistory(String transactionID, String date, float amount, TypeOfTransaction type) {
         this.transactionID = transactionID;
         this.date = date;
         this.amount = amount;
